fix: correct data type for amount column in Statements table

+++ SQL/Statements.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS Statements (
       dp_id VARCHAR(6) not null PRIMARY KEY,
-      amount INT NOT NULL,
+      amount DECIMAL(10,2) NOT NULL,,
       deposit_date datetime DEFAULT NULL
     )
